validator: don’t modify external `props` argument
The `Validator` constructor adds the `props` argument, coming
from the route object itself, to another `Prop` as a child.
This can change the `path` property of the passed `Prop`, which
would create the confusing circumstance that the constructor
modifies its incoming arguments.

To solve this problem, this adds simple deep cloning for
the incoming `opt.props`.

Ref: https://git.help.com/common-backend/nano/issues/23