object: fix allowNull for objects
Previously, an object that was marked allowNull that had child props
would not generate properly due to the allowNull being at the same
level as the object checks. This change wraps all other checks for
objects that allowNull inside a block that is only executed in the
event the object is not null.

Fixes: https://git.help.com/common-backend/help-gen/issues/25
Semver: patch